updated rules configuration
