+++ backend/cmd/users_add.go
@@ -8,7 +8,6 @@ import (
 
 func init() {
 	usersCmd.AddCommand(usersAddCmd)
-	addUserFlags(usersAddCmd.Flags())
 }
 
 var usersAddCmd = &cobra.Command{
