+++ backend/cmd/users_update.go
@@ -8,10 +8,6 @@ import (
 
 func init() {
 	usersCmd.AddCommand(usersUpdateCmd)
-
-	usersUpdateCmd.Flags().StringP("password", "p", "", "new password")
-	usersUpdateCmd.Flags().StringP("username", "u", "", "new username")
-	addUserFlags(usersUpdateCmd.Flags())
 }
 
 var usersUpdateCmd = &cobra.Command{
