+++ backend/settings/config.go
@@ -66,18 +66,18 @@ func setDefaults() Settings {
 				Host: "",
 			},
 		},
-		UserDefaults: UserDefaults{
-			Scope:        ".",
-			LockPassword: false,
-			HideDotfiles: true,
-			Permissions: users.Permissions{
-				Create:   true,
-				Rename:   true,
-				Modify:   true,
-				Delete:   true,
-				Share:    true,
-				Download: true,
-			},
-		},
 	}
+}
+
+// Apply applies the default options to a user.
+func (d *UserDefaults) Apply(u *users.User) {
+	u.Scope = d.Scope
+	u.Locale = d.Locale
+	u.ViewMode = d.ViewMode
+	u.SingleClick = d.SingleClick
+	u.Perm = d.Perm
+	u.Sorting = d.Sorting
+	u.Commands = d.Commands
+	u.HideDotfiles = d.HideDotfiles
+	u.DateFormat = d.DateFormat
 }
