+++ backend/settings/testingConfig.yaml
@@ -47,6 +47,5 @@ userDefaults:
     delete: true
     share: true
     download: true
-  commands: []
   hideDotfiles: false
   dateFormat: false
